Add 3 image sizes (with EventListener) in MovieDetails.jsx
