+++ src/App.jsx
@@ -1,13 +1,13 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import { MovieList } from "./components/MovieList"
-import { MovieDetail } from "./components/MovieDetail"
+import { MovieDetails } from "./components/MovieDetails"
 
 export const App = () => {
   return (
     <BrowserRouter>
       <Routes>
         <Route path="/" element={<MovieList />} />
-        <Route path="/movies/:id" element={<MovieDetail />} />
+        <Route path="/movies/:id" element={<MovieDetails />} />
       </Routes>
     </BrowserRouter>
   );
